moved inner class
